+++ lib/kitchen/directions/bake_learning_objectives.rb
@@ -9,7 +9,10 @@ module Kitchen
         end
       end
 
-      def self.v2(chapter:, add_title: true, li_numbering: :in_chapter)
+      def self.v2(chapter:,
+                  add_title: true,
+                  li_numbering: :in_chapter,
+                  numbering_mode: :chapter_page)
         learning_objectives =
           if %i[in_appendix count_only_li_in_appendix].include?(li_numbering)
             chapter.search('section.learning-objectives')
@@ -30,11 +33,13 @@ module Kitchen
             numbering_type =
               case li_numbering
               when :in_appendix
-                "#{chapter.count_in(:book)}.#{abstract.count_in(:page)}.#{index + 1}"
+                page = chapter
+                "#{page.count_in(:book)}.#{abstract.count_in(:page)}.#{index + 1}"
               when :count_only_li, :count_only_li_in_appendix
                 index + 1
               else
-                "#{chapter.count_in(:book)}.#{abstract.count_in(:chapter)}.#{index + 1}"
+                prefix = abstract.os_number({ mode: numbering_mode, separator: '.' })
+                "#{prefix}.#{index + 1}"
               end
 
             li.replace_children(with: