+++ spec/kitchen_spec/directions/bake_chapter_section_exercises/v1_spec.rb
@@ -81,14 +81,14 @@ RSpec.describe Kitchen::Directions::BakeChapterSectionExercises::V1 do
 
   context 'without deleting title' do
     it 'bakes, keeps in the title & create new one' do
-      described_class.new.bake(chapter: book_with_section_exercises.chapters.first, options: { trash_title: false, create_title: true })
+      described_class.new.bake(chapter: book_with_section_exercises.chapters.first, options: { trash_title: false, create_title: true, numbering_options: {} })
       expect(book_with_section_exercises.chapters.first).to match_snapshot_auto
     end
   end
 
   context 'with title deleted' do
     it 'bakes & removes the title' do
-      described_class.new.bake(chapter: book_with_section_exercises.chapters.first, options: { trash_title: true, create_title: false })
+      described_class.new.bake(chapter: book_with_section_exercises.chapters.first, options: { trash_title: true, create_title: false, numbering_options: {} })
       expect(book_with_section_exercises.chapters.first).to match_snapshot_auto
     end
   end
@@ -97,7 +97,7 @@ RSpec.describe Kitchen::Directions::BakeChapterSectionExercises::V1 do
     let(:extra_title) { '' }
 
     it 'bakes, removes the title & create new one' do
-      described_class.new.bake(chapter: book_with_section_exercises.chapters.first, options: { trash_title: true, create_title: true })
+      described_class.new.bake(chapter: book_with_section_exercises.chapters.first, options: { trash_title: true, create_title: true, numbering_options: {} })
       expect(book_with_section_exercises.chapters.first).to match_snapshot_auto
     end
   end