+++ spec/kitchen_spec/directions/bake_chapter_title/main_spec.rb
@@ -7,4 +7,9 @@ RSpec.describe Kitchen::Directions::BakeChapterTitle do
     expect_any_instance_of(Kitchen::Directions::BakeChapterTitle::V1).to receive(:bake).with(book: 'blah', cases: false)
     described_class.v1(book: 'blah', cases: false)
   end
+
+  it 'calls v2' do
+    expect_any_instance_of(Kitchen::Directions::BakeChapterTitle::V2).to receive(:bake).with(chapters: %w[1 2 3], cases: false, numbering_options: {})
+    described_class.v2(chapters: %w[1 2 3], cases: false, numbering_options: {})
+  end
 end