+++ spec/kitchen_spec/directions/bake_toc_spec.rb
@@ -336,6 +336,15 @@ RSpec.describe Kitchen::Directions::BakeToc do
     expect(book_unit.search('nav').to_s).to match_snapshot_auto
   end
 
+  it 'supports unit numbering' do
+    described_class.v1(
+      book: book_unit,
+      options: { numbering_options: { mode: :unit_chapter_page } }
+    )
+
+    expect(book_unit.search('nav').to_s).to match_snapshot_auto
+  end
+
   it 'works without unit' do
     described_class.v1(book: book_no_unit)
 
@@ -351,7 +360,7 @@ RSpec.describe Kitchen::Directions::BakeToc do
     it 'Page element classes not found' do
       expect {
         described_class.li_for_page(no_pagetype_page)
-      }.to raise_error(RuntimeError, /could not detect which page type class[\s\S]+\(self\) M240:L8:C22/)
+      }.to raise_error(RuntimeError, /could not detect which page type class[\s\S]+\(self\) \/module\/m240\/filename.cnxml:8:22/)
     end
 
     it 'Composite page element classes not found' do